Introduce shared StubHandler fixture; delete orphaned envFactory
